+++ alacritty_terminal/src/term/cell.rs
@@ -1,4 +1,4 @@
-use std::boxed::Box;
+use std::sync::Arc;
 
 use bitflags::bitflags;
 use serde::{Deserialize, Serialize};
@@ -57,19 +57,21 @@ impl ResetDiscriminant<Color> for Cell {
 /// allocation required ahead of time for every cell, with some additional overhead when the extra
 /// storage is actually required.
 #[derive(Serialize, Deserialize, Default, Debug, Clone, Eq, PartialEq)]
-struct CellExtra {
+pub struct CellExtra {
     zerowidth: Vec<char>,
+
+    underline_color: Option<Color>,
 }
 
 /// Content and attributes of a single cell in the terminal grid.
-#[derive(Serialize, Deserialize, Debug, Clone, Eq, PartialEq)]
+#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
 pub struct Cell {
     pub c: char,
     pub fg: Color,
     pub bg: Color,
     pub flags: Flags,
     #[serde(default)]
-    extra: Option<Box<CellExtra>>,
+    pub(crate) extra: Option<Arc<CellExtra>>,
 }
 
 impl Default for Cell {
@@ -94,24 +96,49 @@ impl Cell {
 
     /// Write a new zerowidth character to this cell.
     #[inline]
-    pub fn push_zerowidth(&mut self, c: char) {
-        self.extra.get_or_insert_with(Default::default).zerowidth.push(c);
-    }
+    pub(crate) fn push_zerowidth(&mut self, character: char) {
+        self.init_extra();
 
-    /// Free all dynamically allocated cell storage.
-    #[inline]
-    pub fn drop_extra(&mut self) {
-        if self.extra.is_some() {
-            self.extra = None;
-        }
+        let extra = Arc::make_mut(self.extra.as_mut().unwrap());
+        extra.zerowidth.push(character);
     }
 
     /// Remove all wide char data from a cell.
     #[inline(never)]
-    pub fn clear_wide(&mut self) {
+    pub(crate) fn clear_wide(&mut self) {
         self.flags.remove(Flags::WIDE_CHAR);
-        self.drop_extra();
+        self.extra = None;
         self.c = ' ';
+    }
+
+    /// Set underline color on the cell.
+    pub(crate) fn set_underline_color(&mut self, color: Option<Color>) {
+        // If we reset color and we don't have zerowidth we should drop extra storage.
+        if color.is_none()
+            && self.extra.as_ref().map(|extra| !extra.zerowidth.is_empty()) != Some(false)
+        {
+            self.extra = None;
+            return;
+        }
+
+        self.init_extra();
+
+        let extra = Arc::make_mut(self.extra.as_mut().unwrap());
+        extra.underline_color = color;
+    }
+
+    /// Initializes extra storage if there's none.
+    #[inline]
+    fn init_extra(&mut self) {
+        if self.extra.is_none() {
+            self.extra = Some(Arc::new(Default::default()));
+        }
+    }
+
+    /// Underline color stored in this cell.
+    #[inline]
+    pub fn undreline_color(&self) -> Option<Color> {
+        self.extra.as_ref()?.underline_color
     }
 }
 
@@ -184,10 +211,21 @@ impl LineLength for grid::Row<Cell> {
 
 #[cfg(test)]
 mod tests {
+    use std::mem;
+
     use super::{Cell, LineLength};
 
     use crate::grid::Row;
     use crate::index::Column;
+
+    #[test]
+    fn cell_size_is_below_cap() {
+        // Estimated cell size on 64-bit architectures.
+        const ESTIMATED_CELL_SIZE: usize = 24;
+
+        // Ensure that cell size isn't growning by accident.
+        assert!(mem::size_of::<Cell>() <= ESTIMATED_CELL_SIZE);
+    }
 
     #[test]
     fn line_length_works() {